+++ routes/position/dto/create-position.dto.go
@@ -2,5 +2,5 @@ package dto
 
 type CreatePositionDto struct {
 	Name   string `json:"name" validate:"required"`
-	Salary uint32 `json:"salary" validate:"required,numeric"`
+	Salary uint64 `json:"salary" validate:"required,numeric"`
 }
